% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{accept_package}
\alias{accept_package}
\title{Accept a package on the tracker}
\usage{
accept_package(issue = issue, tarball, note = accept_note(tarball),
  status = 6, ..., session = NULL)
}
\arguments{
\item{issue}{an issue object from \code{\link{issue}}}

\item{note}{The acceptance note to post to the tracker.}

\item{...}{Additional arguments passed to rvest::set_values}

\item{session}{the session to use, if \code{NULL} the issue's session is used.}
}
\description{
Accept a package on the tracker
}
\examples{
\dontrun{
accept_package(1318, "transcriptR_0.99.4.tar.gz")
}
}

