% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{add_data_experiment_packages}
\alias{add_data_experiment_packages}
\title{Add data experiment packages to SVN}
\usage{
add_data_experiment_packages(x, svn_location = "~/proj/experiment",
  manifest = "pkgs/bioc-data-experiment.3.3.manifest")
}
\arguments{
\item{x}{package tarballs to add.}

\item{svn_location}{location of the SVN repository}

\item{manifest}{name of the manifest file}
}
\description{
Add data experiment packages to SVN
}

