% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{add_data_experiment_permisions}
\alias{add_data_experiment_permisions}
\title{Helper function to Add Data Experiment Permissions}
\usage{
add_data_experiment_permisions(x, message = standard_commit_message(x),
  file = "hedgehog:/extra/svndata/gentleman/svn_authz/bioc-data.authz")
}
\arguments{
\item{x}{Permissions to add, can be a named \code{list} or \code{data.frame}.}

\item{message}{Commit message to use}

\item{file}{File containing the permissions to edit}
}
\description{
Helper function to Add Data Experiment Permissions
}

