% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{add_packages}
\alias{add_data_experiment_packages}
\alias{add_packages}
\alias{add_software_packages}
\title{Add packages to SVN}
\usage{
add_software_packages(x, svn_location = "~/proj/Rpacks",
  manifest = "bioc_3.3.manifest")

add_data_experiment_packages(x, svn_location = "~/proj/experiment",
  manifest = "pkgs/bioc-data-experiment.3.3.manifest")
}
\arguments{
\item{x}{package tarballs to add.}

\item{svn_location}{location of the SVN repository}

\item{manifest}{name of the manifest file}
}
\description{
Add packages to SVN
}
\section{Functions}{
\itemize{
\item \code{add_software_packages}: Software Packages

\item \code{add_data_experiment_packages}: Data Experiment Packages
}}

