% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{add_software_packages}
\alias{add_software_packages}
\title{Add software packages to SVN}
\usage{
add_software_packages(x, svn_location = "~/proj/Rpacks",
  manifest = "bioc_3.3.manifest")
}
\arguments{
\item{x}{package tarballs to add.}

\item{svn_location}{location of the SVN repository}

\item{manifest}{name of the manifest file}
}
\description{
Add software packages to SVN
}

