% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{add_to_spreadsheet}
\alias{add_to_spreadsheet}
\title{Add packages to the tracking worksheet}
\usage{
add_to_spreadsheet(x, ..., ws = 1, ss = tracker_spreadsheet())
}
\arguments{
\item{x}{issues to add}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_add_row}}}

\item{ws}{The worksheet number or name to append to.}
}
\description{
This function uses the googlesheets package to add data retrieved from the
tracker.
}
\examples{
\dontrun{
pkgs <- unassigned_packages()
add_to_spreadsheet(pkgs)
}
}

