% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{as.issue}
\alias{as.issue}
\title{Coerce to an issue object}
\usage{
as.issue(x, ...)
}
\arguments{
\item{x}{object to be coerced}

\item{...}{Additional arguments passed to methods}
}
\description{
Coerce to an issue object
}

