% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{assign_new_packages}
\alias{assign_new_packages}
\title{Assign new packages}
\usage{
assign_new_packages(session = tracker_login(), team = devteam,
  pkgs = unassigned_packages(session))
}
\description{
This method uses a hash digest to assign the packages based on the package
name.
}

