% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{assign_new_packages_hash}
\alias{assign_new_packages_hash}
\title{Assign new packages}
\usage{
assign_new_packages_hash(session = tracker_login(), team = devteam,
  pkgs = unassigned_packages(session))
}
\arguments{
\item{session}{the HTTP session to use}

\item{team}{the devteam members to assign}

\item{pkgs}{packages to assign}
}
\description{
This method uses a hash digest to assign the packages based on the package
name.
}

