% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{assign_packages}
\alias{assign_package}
\alias{assign_packages}
\title{Assign a specific package}
\usage{
assign_packages(pkgs, code = assign_new_packages(...), ...)

assign_package(issue, assignee, ...)
}
\arguments{
\item{number}{issue number}

\item{assignment}{lookup assignee by name}
}
\description{
Assign a specific package
}
\section{Functions}{
\itemize{
\item \code{assign_packages}: assign multiple packages with assignment code
}}

