% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_views.R
\name{bioc_views_classification}
\alias{bioc_views_classification}
\title{Extract and validiate biocViews terms from tarball}
\usage{
bioc_views_classification(files)
}
\arguments{
\item{files}{tar.gz file locations}
}
\value{
named vector classifying each tar ball to biocViews hierarchy
}
\description{
Extract and validiate biocViews terms from tarball
}

