% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackage.R
\name{clean}
\alias{clean}
\title{Clean a Software Package}
\usage{
clean(tarball, svnDir = proj_path("Rpacks"), copyToSvnDir = TRUE,
  svnAccountExists = FALSE)
}
\arguments{
\item{tarball}{package tarball}

\item{svnDir}{Directory of the Rpacks checkout}

\item{copyToSvnDir}{whether to copy the files to the SVN directory}
}
\description{
This is for cleaning up build tarballs, and then putting them into
svn (and emailing the authors to let them know this - when they
already have an account)
}

