% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackage.R
\name{clean_data_package}
\alias{clean_data_package}
\title{Clean and copy a Data Experiment package}
\usage{
clean_data_package(tarball, svn_pkgs = "~/proj/experiment/pkgs",
  svn_data_store = "~/proj/experiment/data_store", data_dirs = c("data",
  "inst/extdata"))
}
\arguments{
\item{tarball}{package tarball}

\item{svn_pkgs}{the location of Data Experiment \sQuote{pkgs} checkout.}

\item{svn_data_store}{the location of Data Experiment \sQuote{data_store} checkout.}
}
\value{
File paths to the copied locations (invisibly).
}
\description{
Clean and copy a Data Experiment package
}
\examples{
\dontrun{
pkg <- system.file(package="BiocContributions",
  "testpackages", "RNASeqPower_1.11.0.tar.gz")
clean_data_package(pkg)
}
}

