% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\docType{data}
\name{devteam}
\alias{devteam}
\title{Members of the devteam}
\format{\preformatted{ Named chr [1:7] "Jim Hester" "Martin Morgan" "Valerie Obenchain" "Hervé Pagès" "Dan Tenenbaum" "Brian Long" "Jim Java"
 - attr(*, "names")= chr [1:7] "jhester" "mtmorgan" "vobencha" "herve" ...
}}
\usage{
devteam
}
\description{
Members of the devteam
}
\keyword{datasets}

