% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{edit_data_experiment_permissions}
\alias{edit_data_experiment_permissions}
\alias{edit_data_experiment_permissions.data.frame}
\alias{edit_data_experiment_permissions.list}
\title{Edit the data experiment permissions}
\usage{
edit_data_experiment_permissions(x, ...)

\method{edit_data_experiment_permissions}{data.frame}(x,
  data = read_permissions("hedgehog:/extra/svndata/gentleman/svn_authz/bioc-data.authz"),
  version = 3.2, ...)

\method{edit_data_experiment_permissions}{list}(x,
  data = read_permissions("hedgehog:/extra/svndata/gentleman/svn_authz/bioc-data.authz"),
  version = "3.2")
}
\arguments{
\item{x}{The edits to perform}

\item{...}{Additional arguments passed to methods}

\item{data}{a authz data file}

\item{version}{The release version number}
}
\description{
Edit the data experiment permissions
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: data.frame input, expects columns \sQuote{package} and \sQuote{user}

\item \code{list}: list input, expects a named list of packages and users
}}

