% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{edit_software_permissions}
\alias{edit_software_permissions}
\alias{edit_software_permissions.data.frame}
\alias{edit_software_permissions.list}
\title{Edit the software permissions}
\usage{
edit_software_permissions(x, ...)

\method{edit_software_permissions}{data.frame}(x, data = read_permissions(),
  version = 3.2, ...)

\method{edit_software_permissions}{list}(x, data = read_permissions(),
  version = "3.2")
}
\arguments{
\item{x}{The edits to perform}

\item{...}{Additional arguments passed to methods}

\item{data}{a authz data file}

\item{version}{The release version number}
}
\description{
Edit the software permissions
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: data.frame input, expects columns \sQuote{package} and \sQuote{user}

\item \code{list}: list input, expects a named list of packages and users
}}

