% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{email}
\alias{email}
\alias{email.character}
\alias{email.list}
\alias{email.person}
\alias{email.user_matches}
\title{Extract the email from an object}
\usage{
email(x, ...)

\method{email}{person}(x, ...)

\method{email}{user_matches}(x, ...)

\method{email}{list}(x, ...)

\method{email}{character}(x, ...)
}
\arguments{
\item{x}{the object to extract email from}

\item{...}{Additional arguments passed to methods.}
}
\description{
Extract the email from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{person}: person object

\item \code{user_matches}: user match object from \code{\link{match_user}}

\item \code{list}: list - calls \code{\link{email}} on every item in the list.

\item \code{character}: - simply returns the character vector unchanged.
}}

