% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{emailMaintainer}
\alias{emailMaintainer}
\title{Email a new user their credentials}
\usage{
emailMaintainer(tarball, userId = "user.id", password = "password",
  senderName = getOption("bioc_contributions_signature", "Bioconductor"))
}
\arguments{
\item{tarball}{the package tarball to email about}

\item{userId}{The SVN user ID for the maintainer}

\item{password}{The SVN password for the maintainer}

\item{senderName}{The name of the email sender for use in the signature}
}
\description{
Email a new user their credentials
}

