% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{get_issue}
\alias{get_issue}
\title{Retrieve all of the messages from an issue}
\usage{
get_issue(session = tracker_login(), number)
}
\arguments{
\item{session}{the HTTP session to use}

\item{number}{the issue number to retrieve}
}
\description{
Retrieve all of the messages from an issue
}

