% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_issue}
\alias{gh_issue}
\title{Retrieve all messages associated with an issue}
\usage{
gh_issue(number)
}
\arguments{
\item{number}{The issue number}
}
\value{
a data frame/issue of messages associated with issue \code{number}
}
\description{
Retrieve all messages associated with an issue
}
\examples{
gh_issue(21)
}

