% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_packages_assigned_to}
\alias{gh_packages_assigned_to}
\title{Show packages assigned to a github user}
\usage{
gh_packages_assigned_to(github_username)
}
\arguments{
\item{github_username}{The github username}
}
\value{
a data frame of packages assigned to \code{github_username}
}
\description{
Show packages assigned to a github user
}
\examples{
gh_packages_assigned_to("dtenenba")
}

