% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_pre_accepted_packages}
\alias{gh_pre_accepted_packages}
\title{Show pre-accepted packages}
\usage{
gh_pre_accepted_packages()
}
\value{
a data frame of pre-accepted packages or NULL if none.
}
\description{
Show pre-accepted packages
}
\examples{
gh_pre_accepted_packages()
}

