% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_tracker_search}
\alias{gh_tracker_search}
\title{Query the issue tracker}
\usage{
gh_tracker_search(columns = c("id", "activity", "title", "creator", "status"),
  sort = desc("activity"), filter = c("status", "assignedto"),
  status = c(-1, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10), ...,
  session = tracker_login())
}
\arguments{
\item{columns}{which columns to return}

\item{sort}{A column to sort the data by}

\item{filter}{what columns are used to filter}

\item{status}{the status codes used to filter}

\item{...}{Additional query parameters}

\item{session}{the HTTP session to use}
}
\description{
Query the issue tracker
}
\examples{
tracker_search("@search_text" = "normalize450k")
}

