% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_unassigned_packages}
\alias{gh_unassigned_packages}
\title{Show unassigned packages}
\usage{
gh_unassigned_packages()
}
\value{
a data frame of unassigned packages or NULL if none.
}
\description{
Show unassigned packages
}
\examples{
gh_unassigned_packages()
}

