% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\docType{data}
\name{github_usernames}
\alias{github_usernames}
\title{and their github usernames}
\format{An object of class \code{character} of length 4.}
\usage{
github_usernames
}
\description{
and their github usernames
}
\keyword{datasets}

