% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{issue}
\alias{issue}
\title{Retrieve all of the messages from an issue}
\usage{
issue(number, session = tracker_login())
}
\arguments{
\item{number}{the issue number to retrieve}

\item{session}{the HTTP session to use}
}
\description{
Retrieve all of the messages from an issue
}

