% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{maintainers}
\alias{maintainers}
\title{Retrieve the maintainers from a tarball}
\usage{
maintainers(tarball)
}
\arguments{
\item{tarball}{the package tarball to read}
}
\value{
each maintainer as a 'person' object.
}
\description{
Retrieve the maintainers from a tarball
}
\examples{
pkg <- system.file(package="BiocContributions",
  "testpackages", "RNASeqPower_1.11.0.tar.gz")
maintainers(pkg)
}

