% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{make_github_request}
\alias{make_github_request}
\title{Make a GitHub API request using an OAuth token}
\usage{
make_github_request(uri, method = GET, postdata = NULL)
}
\arguments{
\item{uri}{github uri such as '/orgs/octokit/repos'}

\item{method}{httr verb function (GET, POST, etc)}

\item{postdata}{data to POST (if method == POST)}
}
\value{
an R object created by reading the JSON of the response
}
\description{
Make a GitHub API request using an OAuth token
}

