% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{match_user}
\alias{match_user}
\title{See if a person already exists in the user db}
\usage{
match_user(x)
}
\arguments{
\item{x}{a person object to lookup}
}
\description{
See if a person already exists in the user db
}
\examples{
pkg <- system.file(package="BiocContributions",
  "testpackages", "RNASeqPower_1.11.0.tar.gz")

maintainer <- maintainers(pkg)
match_user(maintainer)
}

