% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackage.R
\name{package_name}
\alias{package_name}
\title{Extract a packages name froma tarball}
\usage{
package_name(tarball)
}
\arguments{
\item{tarball}{package tarball}
}
\value{
the package name
}
\description{
Extract a packages name froma tarball
}
\examples{
pkg <- system.file(package="BiocContributions",
  "testpackages", "RNASeqPower_1.11.0.tar.gz")
package_name(pkg)
}

