% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{post_issue}
\alias{post_issue}
\title{Post a message to an issue}
\usage{
post_issue(issue = NULL, number = NULL, session = NULL, note = edit(),
  file = NULL, ...)
}
\arguments{
\item{issue}{an issue object from \code{\link{get_issue}}}

\item{number}{an issue number, only used if \code{issue} is \code{NULL}}

\item{session}{the session to use, if \code{NULL} the issue's session is used.}

\item{note}{a note to post to the issue, defaults to opening your editor,
but you can also pass a character string.}

\item{file}{a file to attach to the issue, if \code{TRUE} choose a file using
\code{\link{file.choose}}}

\item{...}{Additional arguments passed to rvest::set_values}
}
\description{
Post a message to an issue
}

