% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAproj_path.R
\name{proj_path}
\alias{proj_path}
\title{Path to file (in the unix sense) within the project}
\usage{
proj_path(file = "", check.exists = TRUE)
}
\arguments{
\item{file}{character(1) file or directory name within the project.}

\item{check.exists}{logical(1) indicating whether the file must
exist}
}
\value{
character(1) file path
}
\description{
The path is composed of
\code{getOption("bioc_contributions_project")} or, if NULL, the
home directory "~", preceeding the \code{file} argument.
}

