% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{read_permissions}
\alias{read_permissions}
\title{Read authz permission file}
\usage{

  read_permissions(file = "hedgehog:/extra/svndata/gentleman/svn_authz/bioconductor.authz")
}
\arguments{
\item{file}{location passed to rsync}
}
\description{
Read authz permission file
}

