% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{request_credentials}
\alias{request_credentials}
\title{Request new credentials for users}
\usage{
request_credentials(x, sender = getOption("bioc_contributions_signature"))
}
\arguments{
\item{x}{an object with an \code{\link{email}} method defined.}

\item{sender}{The name to use in the signature}
}
\value{
A \code{\link[gmailr]{mime}} object.
}
\description{
Request new credentials for users
}
\examples{
pkg <- system.file(package="BiocContributions",
  "testpackages", "RNASeqPower_1.11.0.tar.gz")
request_credentials(match_user(maintainers(pkg)))
request_credentials("asdf@asd.com")
}

