% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{run_commands}
\alias{rcs_check_in}
\alias{rcs_check_out}
\alias{run_commands}
\title{Run commands on a file, possibly remote.}
\usage{

  rcs_check_out(file = "hedgehog:/extra/svndata/gentleman/svn_authz/bioconductor.authz",
  args = NULL)


  rcs_check_in(file = "hedgehog:/extra/svndata/gentleman/svn_authz/bioconductor.authz",
  message, args = NULL)
}
\arguments{
\item{args}{Additional arguments passed to the command.}

\item{message}{commit message for the check in}

\item{inheritParams}{read_permissions}
}
\description{
Run commands on a file, possibly remote.
}
\details{
if the file is a remote location (server:path) the command is run
remotely.
}
\section{Functions}{
\itemize{
\item \code{rcs_check_out}: Check out a RCS tracked file

\item \code{rcs_check_in}: Check in a RCS tracked file
}}

