% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{standard_commit_message}
\alias{standard_commit_message}
\title{Generate a standard commit message for permission edits}
\usage{
standard_commit_message(x)
}
\arguments{
\item{x}{the edits to make, if a data.frame will be coerced to a named list.}
}
\description{
Generate a standard commit message for permission edits
}

