% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\docType{data}
\name{status_map}
\alias{status_map}
\title{Status code to name mapping}
\format{An object of class \code{character} of length 10.}
\usage{
status_map
}
\description{
Status code to name mapping
}
\keyword{datasets}

