% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{svn}
\alias{svn}
\title{Declare an SVN instance}
\usage{
svn(dir = getwd())
}
\arguments{
\item{dir}{The SVN directory}
}
\description{
This is primarily a helper function for using svn progromatically.
}
\examples{
\dontrun{
s <- svn("my/svn/location")
s$status()
s$update()
s$add("file3")
s$commit("adding file3")
}
}

