% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn_auth_text.R
\name{svn_auth_text}
\alias{svn_auth_text}
\title{Generate text to copy / paste into svn file Bioconductor.authz}
\usage{
svn_auth_text(filenames, version = getOption("BC_auth_version", "3_2"))
}
\arguments{
\item{filenames}{character() vector of (full) paths to package tar balls.}

\item{version}{character(1) version string as it appears in svn
'RELEASE' branch, e.g., \code{"3_2"} corresponding to
RELEASE_3_2.}
}
\description{
Generate text to copy / paste into svn file Bioconductor.authz
}

