% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackerCode.R
\name{tabulate_activity}
\alias{tabulate_activity}
\title{Tabulate tracker activity}
\usage{
tabulate_activity(issues = tracker_search(columns = c("id", "activity",
  "title", "creator", "status", "assignedto", "actor")), date = Sys.Date() -
  30, users = devteam)
}
\arguments{
\item{users}{people to tabulate about, default is the devteam.}

\item{Date}{Date to tabulate from, default is the last thirty days.}
}
\description{
Tabulate tracker activity
}

