% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{tracker_login}
\alias{tracker_login}
\title{Login to the issue tracker}
\usage{
tracker_login(url = "https://tracker.bioconductor.org",
  user = getOption("tracker_user"),
  password = getOption("tracker_password"))
}
\arguments{
\item{url}{tracker url}

\item{user}{username used to login}

\item{password}{password used to login}
}
\description{
Login to the issue tracker
}

