% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{tracker_search}
\alias{tracker_search}
\alias{unassigned_packages}
\title{Query the issue tracker}
\usage{
tracker_search(session = tracker_login(), columns = c("id", "activity",
  "title", "creator", "status"), sort = desc("activity"),
  filter = c("status", "assignedto"), status = c(-1, 1, 2, 3, 4, 5, 6, 7, 8,
  9, 10), ...)

unassigned_packages(session = tracker_login(), status = c(-1, 1), ...)
}
\arguments{
\item{session}{the HTTP session to use}

\item{columns}{which columns to return}

\item{sort}{A column to sort the data by}

\item{filter}{what columns are used to filter}

\item{status}{the status codes used to filter}

\item{...}{Additional query parameters}
}
\description{
Query the issue tracker
}
\section{Functions}{
\itemize{
\item \code{unassigned_packages}: retrieve unassigned packages
}}

