% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{user_db}
\alias{user_db}
\title{Retrieve the remote user database}
\usage{
user_db()
}
\value{
the result is a data.frame of the data
}
\description{
This is an internal function which retrieves the user database file
'user_db.csv', which holds user information for users with SVN
credentials.  this assumes 'rsync' is available on your path, which
is true by default for linux and OSX machines, but probably not for
windows.
}
\details{
The information is cached so calling this function repeatably will
result in the same information being returned. Use
\code{memoise::forget(user_db)} to reset the cache if needed.
}
\examples{
user_db()
# second call will be nearly instantaneous
user_db()
# clear the cache
memoise::forget(user_db)
user_db()
}

