% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{users}
\alias{users}
\title{Retrieve the user list}
\usage{
users(session = tracker_login(), ...)
}
\arguments{
\item{session}{the HTTP session to use}

\item{...}{Additional query parameters}
}
\description{
Retrieve the user list
}

