% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_standard.R
\name{workflow_standard}
\alias{CreatePackageAssignmentEmail}
\alias{DownloadNewPackageTarballs}
\alias{DraftWeeklySummaryEmail}
\alias{ManageNewPackagesCredentials}
\alias{workflow_standard}
\title{Workflow steps}
\usage{
CreatePackageAssignmentEmail(assignInTracker = FALSE,
  secret = proj_path("bioconductorseattle-gmail.json"))

DownloadNewPackageTarballs(pre = pre_accepted_packages())

ManageNewPackagesCredentials(metadata, createDraft = TRUE)

DraftWeeklySummaryEmail()
}
\arguments{
\item{assignInTracker}{logical(1) indicating whether package
assignments should be updated in the tracker.}

\item{secret}{character(1) path to JSON API client access secret.}

\item{pre}{data.frame() returned by \code{pre_accepted_packages()},
retrieved from the tracker and indicating packages tagged as
'pre-accepted'}

\item{metadata}{Return value from
\code{DownloadNewPackageTarballs}. If missing, search
\code{proj_path()} for most recent saved version as RData with
format "new-packages-metadata_20160211.RData"}

\item{createDraft}{logical(1) draft email to FHCRC for SVN new user
credentials}
}
\description{
Workflow steps
}

