% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{write_permissions}
\alias{write_permissions}
\title{Write authz permission file}
\usage{
write_permissions(x,
  file = "hedgehog:/extra/svndata/gentleman/svn_authz/bioconductor.authz",
  ...)
}
\arguments{
\item{x}{object to write}

\item{file}{location passed to rsync}
}
\description{
Write authz permission file
}

