% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn.R
\name{add_packages}
\alias{add_packages}
\title{Add packages to SVN}
\arguments{
\item{x}{package tarballs to add.}

\item{svn_location}{location of the SVN repository}

\item{manifest}{name of the manifest file}
}
\description{
Add packages to SVN
}

