% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{assign_new_packages}
\alias{assign_new_packages}
\title{Assign new packages}
\usage{
assign_new_packages(session = tracker_login(), team = names(devteam),
  date = Sys.Date(), pkgs = unassigned_packages(session))
}
\arguments{
\item{session}{the HTTP session to use}

\item{team}{the devteam members to assign}

\item{date}{the date of assignment}

\item{pkgs}{packages to assign}
}
\description{
This method uses \code{sample} to assign the packages and sets the random
seed based on the date.
}

