% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{assign_new_packages}
\alias{assign_new_packages}
\title{Assign new packages}
\usage{
assign_new_packages(pkgs = unassigned_packages(session), team = devteam)
}
\arguments{
\item{pkgs}{the packages to assign}

\item{team}{team members to assign to}
}
\description{
This method uses a hash digest to assign the packages based on the package
name.
}

