% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\docType{data}
\name{devteam}
\alias{devteam}
\title{Members of the devteam}
\format{An object of class \code{character} of length 4.}
\usage{
devteam
}
\description{
Members of the devteam
}
\keyword{datasets}

