% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\docType{data}
\name{devteam}
\alias{devteam}
\title{Members of the devteam}
\format{\preformatted{ Named chr [1:5] "jhester" "mtmorgan" "vobencha" "herve" "dtenenba"
 - attr(*, "names")= chr [1:5] "Jim Hester" "Martin Morgan" "Valerie Obenchain" "Hervé Pagès" ...
}}
\usage{
devteam
}
\description{
Members of the devteam
}
\keyword{datasets}

