% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{download}
\alias{download}
\title{Download attachments from an issue}
\usage{
download(session = tracker_login(), issue, dir = ".", ...)
}
\arguments{
\item{session}{the HTTP session to use}

\item{issue}{The issue object, or issue number to download files from}

\item{dir}{the location to store the files}

\item{...}{Additional query parameters}
}
\description{
Download attachments from an issue
}

