% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{make_github_request}
\alias{make_github_request}
\title{Make a GitHub API request using an OAuth token}
\usage{
make_github_request(uri, method = GET, postdata = NULL,
  include_message = FALSE)
}
\arguments{
\item{uri}{github uri such as '/orgs/octokit/repos'}

\item{method}{httr verb function (GET, POST, etc)}

\item{postdata}{data to POST (if method == POST)}

\item{include_message}{Whether to include the message in the result}
}
\value{
a data frame based on the result from github
}
\description{
Make a GitHub API request using an OAuth token
}

