% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svn_auth_text.R
\name{svn_auth_text}
\alias{svn_auth_text}
\alias{svn_data_experiment_auth_text}
\alias{svn_software_auth_text}
\title{Update svn authz files}
\usage{
svn_software_auth_text(filenames, version = getOption("BC_auth_version",
  "3_3"))

svn_data_experiment_auth_text(filenames,
  version = getOption("BC_auth_version", "3_3"))
}
\arguments{
\item{filenames}{character() vector of (full) paths to package tar balls.}

\item{version}{character(1) version string as it appears in svn
'RELEASE' branch, e.g., \code{"3_2"} corresponding to
RELEASE_3_2.}
}
\description{
These functions generate text to copy / paste into software svn
file bioconductor.authz or data experiment svn file
bioc-data.authz
}

