% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{tracker_search}
\alias{my_issues}
\alias{pre_accepted_packages}
\alias{tracker_search}
\alias{unassigned_packages}
\title{Query the issue tracker}
\usage{
tracker_search(columns = c("id", "activity", "title", "creator", "status"),
  sort = desc("activity"), filter = c("status", "assignedto"),
  status = c(-1, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10), ...,
  session = tracker_login())

unassigned_packages(status = c(-1, 1, 2, 3, 4, 5, 9), ...,
  session = tracker_login())

pre_accepted_packages(status = 9, ..., session = tracker_login())

my_issues(user = NULL, status = c(-1, 1, 2, 3, 4, 5, 9, 10), ...,
  session = tracker_login())
}
\arguments{
\item{columns}{which columns to return}

\item{sort}{A column to sort the data by}

\item{filter}{what columns are used to filter}

\item{status}{the status codes used to filter}

\item{...}{Additional query parameters}

\item{session}{the HTTP session to use}
}
\description{
Query the issue tracker
}
\section{Functions}{
\itemize{
\item \code{unassigned_packages}: retrieve unassigned packages

\item \code{pre_accepted_packages}: retrieve pre-accepted packages

\item \code{my_issues}: retrieve the logged in users packages
}}

